requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy==1.2.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
PySocks==1.7.1
python-telegram-bot==20.2
rfc3986==1.5.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
